+++ Lib/openbakery/cli.py
@@ -1,14 +1,12 @@
 import argparse
-import pkgutil
-import runpy
 import signal
 import sys
 from importlib import import_module
 
-import openbakery.commands
+from openbakery import __version__
 from openbakery.commands.check_profile import main as check_profile_main
 
-CLI_PROFILES = [
+SUBCOMMANDS = [
     "adobefonts",
     "fontbureau",
     "fontval",
@@ -17,14 +15,15 @@ CLI_PROFILES = [
     "iso15008",
     "notofonts",
     "opentype",
-    "ufo_sources",
+    "ufo-sources",
     "universal",
     "proposals",
+    "check-profile",
 ]
 
 
 def run_profile_check(profilename):
-    module = import_module(f"openbakery.profiles.{profilename}")
+    module = import_module(f"openbakery.profiles.{profilename.replace('-', '_')}")
     sys.exit(check_profile_main(module.profile))
 
 
@@ -36,54 +35,42 @@ def signal_handler(sig, frame):
 def main():
     signal.signal(signal.SIGINT, signal_handler)
 
-    subcommands = [
-        pkg[1] for pkg in pkgutil.walk_packages(openbakery.commands.__path__)
-    ] + ["check_" + prof for prof in CLI_PROFILES]
-
-    subcommands = [command.replace("_", "-") for command in sorted(subcommands)]
-
-    if len(sys.argv) >= 2 and sys.argv[1] in subcommands:
-        # Relay to subcommand.
+    if len(sys.argv) >= 2 and sys.argv[1] in SUBCOMMANDS:
         subcommand = sys.argv[1]
-        subcommand_module = subcommand.replace("-", "_")
+
         sys.argv[0] += " " + subcommand
         del sys.argv[1]  # Make this indirection less visible for subcommands.
-        if (
-            subcommand_module.startswith("check_")
-            and subcommand_module[6:] in CLI_PROFILES
-        ):
-            run_profile_check(subcommand_module[6:])
+
+        if subcommand == "check-profile":
+            check_profile_main()
         else:
-            runpy.run_module(
-                "openbakery.commands." + subcommand_module, run_name="__main__"
-            )
+            run_profile_check(subcommand)
+
     else:
         description = (
-            "Run openbakery subcommands. Subcommands have their own help "
-            "messages. These are usually accessible with the -h/--help flag "
-            "positioned after the subcommand, i.e.: openbakery subcommand -h"
+            "Run openbakery subcommands. Subcommands have their own help messages;\n"
+            "to view them add the '-h' (or '--help') option after the subcommand,\n"
+            "like in this example:\n    openbakery universal -h"
         )
 
-        parser = argparse.ArgumentParser(description=description)
+        parser = argparse.ArgumentParser(
+            formatter_class=argparse.RawTextHelpFormatter, description=description
+        )
         parser.add_argument(
             "subcommand",
             help="The subcommand to execute",
             nargs="?",
-            choices=subcommands,
+            choices=SUBCOMMANDS,
         )
         parser.add_argument(
             "--list-subcommands",
             action="store_true",
-            help="print the list of subcommands "
-            "to stdout, separated by a space character. This is "
-            "usually only used to generate the shell completion code.",
+            help="print list of supported subcommands",
         )
-        parser.add_argument(
-            "--version", action="version", version=openbakery.__version__
-        )
+        parser.add_argument("--version", action="version", version=__version__)
         args = parser.parse_args()
 
         if args.list_subcommands:
-            print(" ".join(subcommands))
+            print("\n".join(SUBCOMMANDS))
         else:
             parser.print_help()